[refactor] #129 - SettingsViewController, FriendListViewController DB 접근 메서드 FirebaseManager로 이동
